Modernize game/frontend/rendering casts and macros
Squashes 54 original commits from 57ee8701..d33fcb75.
First: Modernize Unix thread and socket casts
Last:  Expand application forward declaration macros

Representative commits:
- 57ee8701 Modernize Unix thread and socket casts
- 91dd4620 Use nullptr for file handles
- 5150c624 Modernize Perlin allocations
- 970df308 Replace strong typedef macros
- 7f1a7f24 Replace exception declaration macro
- d33fcb75 Expand application forward declaration macros

+++ source/game/StarCollisionBlock.hpp
@@ -102,7 +102,7 @@ inline bool CollisionSet::contains(CollisionKind kind) const {
 }
 
 inline uint8_t CollisionSet::kindBit(CollisionKind kind) {
-  return 1 << ((uint8_t)kind + 1);
+  return static_cast<uint8_t>(1u << (static_cast<uint8_t>(kind) + 1));
 }
 
 inline bool isColliding(CollisionKind kind, CollisionSet const& collisionSet) {
